# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "实验室人脸识别打卡系统",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/kiosk/device": {
      "get": {
        "tags": [
          "打卡终端"
        ],
        "summary": "Device Status",
        "operationId": "device_status_api_v1_kiosk_device_get",
        "parameters": [
          {
            "name": "X-Device-Code",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Device-Code"
            }
          },
          {
            "name": "X-Device-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Device-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevicePublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/kiosk/dashboard": {
      "get": {
        "tags": [
          "打卡终端"
        ],
        "summary": "Kiosk Dashboard",
        "operationId": "kiosk_dashboard_api_v1_kiosk_dashboard_get",
        "parameters": [
          {
            "name": "X-Device-Code",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Device-Code"
            }
          },
          {
            "name": "X-Device-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Device-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KioskDashboard"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/kiosk/presence": {
      "get": {
        "tags": [
          "打卡终端"
        ],
        "summary": "Kiosk Presence",
        "operationId": "kiosk_presence_api_v1_kiosk_presence_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          },
          {
            "name": "X-Device-Code",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Device-Code"
            }
          },
          {
            "name": "X-Device-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Device-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KioskPresencePage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/kiosk/records": {
      "get": {
        "tags": [
          "打卡终端"
        ],
        "summary": "Kiosk Records",
        "operationId": "kiosk_records_api_v1_kiosk_records_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "X-Device-Code",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Device-Code"
            }
          },
          {
            "name": "X-Device-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Device-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KioskRecordPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/kiosk/recognition-sessions": {
      "post": {
        "tags": [
          "打卡终端"
        ],
        "summary": "Create Recognition Session",
        "operationId": "create_recognition_session_api_v1_kiosk_recognition_sessions_post",
        "parameters": [
          {
            "name": "X-Device-Code",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Device-Code"
            }
          },
          {
            "name": "X-Device-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Device-Key"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecognitionSessionPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/kiosk/recognition-sessions/{session_id}/verify": {
      "post": {
        "tags": [
          "打卡终端"
        ],
        "summary": "Verify Recognition Session",
        "operationId": "verify_recognition_session_api_v1_kiosk_recognition_sessions__session_id__verify_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          },
          {
            "name": "X-Device-Code",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Device-Code"
            }
          },
          {
            "name": "X-Device-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Device-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_verify_recognition_session_api_v1_kiosk_recognition_sessions__session_id__verify_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecognitionVerifyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/kiosk/attendance-actions": {
      "post": {
        "tags": [
          "打卡终端"
        ],
        "summary": "Attendance Action",
        "operationId": "attendance_action_api_v1_kiosk_attendance_actions_post",
        "parameters": [
          {
            "name": "X-Device-Code",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Device-Code"
            }
          },
          {
            "name": "X-Device-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Device-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttendanceActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendancePublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttendanceActionRequest": {
        "properties": {
          "ticket": {
            "type": "string",
            "title": "Ticket"
          },
          "action": {
            "$ref": "#/components/schemas/AllowedAction"
          },
          "idempotency_key": {
            "type": "string",
            "maxLength": 64,
            "minLength": 8,
            "title": "Idempotency Key"
          }
        },
        "type": "object",
        "required": [
          "ticket",
          "action",
          "idempotency_key"
        ],
        "title": "AttendanceActionRequest"
      },
...
      "Body_verify_recognition_session_api_v1_kiosk_recognition_sessions__session_id__verify_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_verify_recognition_session_api_v1_kiosk_recognition_sessions__session_id__verify_post"
      },
...
      "DevicePublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "last_seen_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "name",
          "location",
          "is_active",
          "last_seen_at",
          "created_at"
        ],
        "title": "DevicePublic"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KioskDashboard": {
        "properties": {
          "device": {
            "$ref": "#/components/schemas/DevicePublic"
          },
          "engine_status": {
            "type": "string",
            "title": "Engine Status"
          },
          "engine_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine Version"
          },
          "server_time": {
            "type": "string",
            "format": "date-time",
            "title": "Server Time"
          },
          "current_count": {
            "type": "integer",
            "title": "Current Count"
          },
          "today_checkins": {
            "type": "integer",
            "title": "Today Checkins"
          },
          "today_checkouts": {
            "type": "integer",
            "title": "Today Checkouts"
          },
          "exception_count": {
            "type": "integer",
            "title": "Exception Count"
          },
          "recent_records": {
            "items": {
              "$ref": "#/components/schemas/KioskRecordItem"
            },
            "type": "array",
            "title": "Recent Records"
          }
        },
        "type": "object",
        "required": [
          "device",
          "engine_status",
          "server_time",
          "current_count",
          "today_checkins",
          "today_checkouts",
          "exception_count",
          "recent_records"
        ],
        "title": "KioskDashboard"
      },
      "KioskPresencePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CurrentUserItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "KioskPresencePage"
      },
...
      "KioskRecordPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/KioskRecordItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "KioskRecordPage"
      },
...
      "RecognitionSessionPublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "challenge": {
            "$ref": "#/components/schemas/ChallengeType"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "challenge",
          "prompt",
          "expires_at"
        ],
        "title": "RecognitionSessionPublic"
      },
      "RecognitionVerifyResponse": {
        "properties": {
          "recognized": {
            "type": "boolean",
            "title": "Recognized"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "real_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Real Name"
          },
          "allowed_action": {
            "$ref": "#/components/schemas/AllowedAction",
            "default": "BLOCKED"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "ticket": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ticket"
          },
          "match_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Score"
          },
          "liveness_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Liveness Score"
          },
          "face_box": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FaceBox"
              },
              {
                "type": "null"
              }
            ]
          },
          "quality_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Hint"
          },
          "processing_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Ms"
          },
          "processed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processed At"
          }
        },
        "type": "object",
        "required": [
          "recognized",
          "message"
        ],
        "title": "RecognitionVerifyResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}